requirements.txt
browser-use==0.11.9
python-dotenv>=1.0.1
playwright
fastapi>=0.115
uvicorn[standard]>=0.34
jinja2>=3.1
pyyaml>=6.0
aiosqlite>=0.20
